Flyttade delar av controllern till en ny inputhandler-klass
