+++ src/main/java/TestGrupp/Controller/Controller.java
@@ -4,13 +4,16 @@ import TestGrupp.Model.GameModel;
 import java.awt.*;
 import java.awt.event.*;
 import TestGrupp.View.Panel;
-import TestGrupp.Model.GameModel;
+
 import javax.swing.*;
 
-public class Controller implements KeyListener{
+
+
+public class Controller {
     //a basic controller that handles KEYBOARD input
 
     GameModel gm;
+    InputHandler ih;
     Panel panel;
 
     Action upAction;
@@ -20,39 +23,15 @@ public class Controller implements KeyListener{
 
         this.gm = gm;
         this.panel = panel;
-        this.panel.addKeyListener(this);
+        this.ih = new InputHandler(this.panel);
+
 
         this.panel.setFocusable(true);
         this.panel.requestFocusInWindow();
 
    }
 
-    @Override
-    public void keyTyped(KeyEvent e) {
 
-    }
-
-    @Override
-    public void keyPressed(KeyEvent e) {
-        int key = e.getKeyCode();
-        if (key == KeyEvent.VK_LEFT) {
-            panel.getSpaceShipView().setLocation(panel.getSpaceShipView().getX() - 10, panel.getSpaceShipView().getY());
-        }
-        if (key == KeyEvent.VK_RIGHT) {
-            panel.getSpaceShipView().setLocation(panel.getSpaceShipView().getX() + 10, panel.getSpaceShipView().getY());
-        }
-        if (key == KeyEvent.VK_UP) {
-            panel.getSpaceShipView().setLocation(panel.getSpaceShipView().getX(), panel.getSpaceShipView().getY() - 10);
-        }
-        if (key == KeyEvent.VK_DOWN) {
-            panel.getSpaceShipView().setLocation(panel.getSpaceShipView().getX(), panel.getSpaceShipView().getY() + 10);
-        }
-    }
-
-    @Override
-    public void keyReleased(KeyEvent e) {
-
-    }
 
     public void update() {
         panel.repaint();
